requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow-cpu==2.15.0
pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
streamlit==1.29.0
requests==2.31.0
tensorflow-datasets==4.9.3
locust==2.19.1
gdown==4.7.1
